stata_harvester/pyproject.toml
[project]
name = "stata-harvester"
version = "0.1.0"
description = "Add your description here"
readme = "README.md"
requires-python = ">=3.12"
dependencies = [
    "apache-airflow>=2.10.5",
    "common",
    "pandas>=2.2.3",
]

[tool.uv.sources]
common = { git = "https://github.com/opendatabs/common", rev = "v0.1.0" }

[dependency-groups]
dev = [
    "ruff",
]
